Introduce custom executable id (#46)
Before this commit the only unique identifier for an executable file or
object file was the build id. As the GNU's build ID might not be present,
we fall back to the sha256 hash of the `.text` section. This works fine but
takes quite a bit of space as we refer to it in every mapping.

By using a 64 bit id derived from the first 8 bytes of the sha256 hash
of the code, we can save some memory, make comparisons and hash the IDs faster,
and helps simplify the code. For example, there's no need for a separate ID in
the native unwind state. There might be a bit more CPU and IO needed to hash
the code, but this might be something we can cache in the future.

This is also useful for future profile formats that require unique 64 bit
IDs for object files.

Also fixes other minor things, such as using the lower hexadecimal representation.

Test Plan
=========

Lots of manual tests, everything seems good.

+++ src/bpf/profiler.bpf.c
@@ -52,7 +52,7 @@ struct {
 struct {
   __uint(type, BPF_MAP_TYPE_HASH);
   __uint(max_entries, 5 * 1000);
-  __type(key, u32);
+  __type(key, u64);
   __type(value, unwind_info_chunks_t);
 } unwind_info_chunks SEC(".maps");
 
@@ -116,9 +116,9 @@ static __always_inline u64 find_offset_for_pc(stack_unwind_table_t *table, u64 p
 // address.
 static __always_inline chunk_info_t*
 find_chunk(mapping_t *mapping, u64 object_relative_pc) {
-  u32 executable_id = mapping->executable_id;
+  u64 executable_id = mapping->executable_id;
 
-  LOG("~about to check chunks, executable_id=%d", executable_id);
+  LOG("~about to check chunks, executable_id=%lld", executable_id);
   // Find the chunk where this unwind table lives.
   // Each chunk maps to exactly one shard.
   unwind_info_chunks_t *chunks =
